Solve part 2 of Day 01
So this involves a bit of a silly hack.

Problem - The code for finding the last digit of a line was not working.
Solution - Rewrite the code so that it finds the first digit of a
reversed copy of the string, using the code I know works.

Also, 3 spelled backwards is "eerht" for future reference.