Refactor Day 03
So most of the code that cares about the grid primarily cares about the
numbers in the grid, so every function that took the grid immediately
converted it into grid numbers. So I factored that part out and now
part one can work exactly like it did before I knew what part two was
going to ask of me. Huzzah!