Solve part 2 of Day 08
So I used a common hack for this: the A to Z paths all cycle, so all I
need to do is find a least common multiple. I'd love to find a solution
that doesn't rely on an unspoken rule of the input, but that will have
to come at a later date.