Solve part 2 of Day 09
That was ...weird. Calculating the previous value manually worked on the
test data, but it provided a completely different and NEGATIVE result on
my input data. I'm not quite sure why, but somebody pointed out a useful
trick that did give me an answer. A mystery for a later time.